+++ app/src/main/java/app/morphe/manager/patcher/runtime/CoroutineRuntime.kt
@@ -40,14 +40,13 @@ class CoroutineRuntime(private val context: Context) : Runtime(context) {
                     .filterKeys { it in selectedBundles }
 
             val patchList = selectedPatches.flatMap { (bundle, selected) ->
-                allPatches[bundle]?.filter { it.name in selected }
+                allPatches[bundle]?.filterKeys { it in selected }?.values
                     ?: throw IllegalArgumentException("Patch bundle $bundle does not exist")
             }
 
             // Set all patch options.
             options.forEach { (bundle, bundlePatchOptions) ->
-                val patches = allPatches[bundle] ?: return@forEach
-                val patchesByName = patches.associateBy { it.name }
+                val patchesByName = allPatches[bundle] ?: return@forEach
 
                 bundlePatchOptions.forEach { (patchName, configuredPatchOptions) ->
                     // Morphe: Skip if patch doesn't exist in this bundle